% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{apply_multi_micro_clustering}
\alias{apply_multi_micro_clustering}
\title{Perform microclustering on dataset for pseudocell generation}
\usage{
apply_multi_micro_clustering(
  seurat,
  cells_per_partition = 100,
  find_neighbors = FALSE,
  reduction1 = "harmony",
  reduction2 = "harmony_peaks",
  umap_name = "WNN.UMAP",
  assay_name = "SCT",
  neighbor_name = "weighted.nn",
  dims = list(1:50, 1:50),
  k.nn = 5
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{cells_per_partition}{Targeted number of cells per partition}

\item{find_neighbors}{Boolean designating whether to run Seurat's FindMultiModalNeighbors. Set to FALSE if this has already been run}

\item{reduction1}{Reduction 1 to use for FindMultiModalNeighbors, if find_neighbors is TRUE}

\item{reduction2}{Reduction 2 to use for FindMultiModalNeighbors, if find_neighbors is TRUE}

\item{k.nn}{Number of nearest neighbors to use for KNN algorithm, if find_neighbors is TRUE}

\item{dim_list}{List of dimensions to use for FindMultiModalNeighbors, if find_neighbors is TRUE}
}
\value{
List of individual cell communities that make up each called pseudocell

List of individual cell barcodes for each pseudocell
}
\description{
This function takes as input a Seurat object and generates a
weighted nearest neighbor (WNN) graph if needed. It then performs
commnunity detection with Louvain clustering and returns a list of pools
denoting the individual cells that comprise each pseudocell
}
