% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{conns_to_peak_gene_key}
\alias{conns_to_peak_gene_key}
\title{Utility function used to generate peak-gene key dataframe from the output of Cicero or an equivalent method}
\usage{
conns_to_peak_gene_key(conns, seurat_atac, mc.cores = 8)
}
\arguments{
\item{conns}{Connection dataframe generated by Cicero or equivalent method}

\item{seurat_atac}{Single cell ATAC-seq Seurat object with stored Annotation}

\item{mc.cores}{Number of threads to use for multithreading}
}
\value{
Peak-gene dataframe for use by run_tf_aen_ccan
}
\description{
This function takes in a conns dataframe generated by Cicero or an equivalent method and
a single cell ATAC-seq Seurat object. It generates CCANs then uses the Annotation
stored in the Seurat object to identify CCANs with peaks that overlap with a gene's position
to link putative cis-regulatory elements to each gene.
}
