% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{make_aen_lists}
\alias{make_aen_lists}
\title{Utility function that extracts names of peaks with predicted non-zero regulatory effect}
\usage{
make_aen_lists(peak_results)
}
\arguments{
\item{peak_results}{Output of run_peak_aen, in list of lists format}
}
\value{
List of character vectors, corresponding to predicted CRE names for each modeled gene
}
\description{
This function takes in the output of run_peak_aen and returns
names of predicted CREs, by non-zero regulatory effect, for each
modeled gene.
}
