% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{make_bootstrap_sequences}
\alias{make_bootstrap_sequences}
\title{Generate seeds for reproducibility when sampling from the starting dataset to perform bootstrapping}
\usage{
make_bootstrap_sequences(num_bootstraps, gene_names)
}
\arguments{
\item{num_bootstraps}{Number of bootstrap samples}

\item{gene_names}{Modeled gene names}

\item{dims.list}{A list containing the dimensions for each reduction to use}
}
\value{
List named by inputted gene names, of lists of seeds, each of length num_bootstraps
}
\description{
This function will generate seeds for bootstrapping for each modeled gene.
It takes in the desired number of bootstrap samples and the list of genes as input
and returns a list, named by inputted genes, of lists of seeds.
}
