% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{make_cre_model_df}
\alias{make_cre_model_df}
\title{Utility function used to convert list of lists output of run_peak_aen to more useful dataframe format}
\usage{
make_cre_model_df(aen_results_peaks)
}
\arguments{
\item{aen_results_peaks}{Modeling output of run_peak_aen}
}
\value{
CRE-gene dataframe containing modeling results, with rows representing each modeled gene.
}
\description{
This function converts the output of run_peak_aen to a CRE by gene dataframe.
Each position is the predicted regulatory weight of the given CRE for the
given gene's expression.
}
