% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{make_motif_matrix}
\alias{make_motif_matrix}
\title{Utility function used to generate}
\usage{
make_motif_matrix(
  seurat_atac,
  pwm_list,
  ref_genome = BSgenome.Hsapiens.UCSC.hg38
)
}
\arguments{
\item{seurat_atac}{Single cell ATAC-seq Seurat object with stored Annotation}

\item{pwm_list}{PWMList to use. Default is human_pwms_v2 from chromVARmotifs}

\item{ref_genome}{BSgenome for AddMotifs to use}
}
\value{
Peak-TF dataframe for use by run_tf_aen_ccan
}
\description{
This function takes in a single cell ATAC-seq Seurat object and a Position Weight Matrix list.
It generates a peak by TF dataframe through Signac's AddMotifs function.
}
