% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{make_tf_model_df}
\alias{make_tf_model_df}
\title{Utility function used to convert list of lists output of run_tf_aen to more useful dataframe format}
\usage{
make_tf_model_df(aen_results)
}
\arguments{
\item{aen_results}{Modeling output of run_tf_aen}
}
\value{
TF-gene dataframe containing modeling results, with rows representing each modeled gene.
}
\description{
This function converts the output of run_tf_aen to a TF by gene dataframe.
Each position is the predicted regulatory weight of the given TF for the
given gene's expression.
}
