% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{peaks_to_tfs_ccan}
\alias{peaks_to_tfs_ccan}
\title{Utility function used to convert a peak list for a given gene to a list of TFs with predicted motifs within the peak list, used in run_tf_aen_ccan}
\usage{
peaks_to_tfs_ccan(peak_list, key_tibble)
}
\arguments{
\item{peak_list}{Vector of peak names}

\item{key_tibble}{Tibble of peaks by motifs denoting presence of each motif in each peak}
}
\value{
Names of TFs with motifs within inputted peaks
}
\description{
This function takes in a peak list and peak-TF key as input and
returns the TF names with at least one motif present in the
peak list, as determined by the peak-TF key.
}
