% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_footprinting}
\alias{plot_footprinting}
\title{Plot Tn5 integration enrichment between two sets of CREs/peaks}
\usage{
plot_footprinting(
  object,
  assay.1,
  assay.2,
  features,
  group.by = NULL,
  idents = NULL,
  label = TRUE,
  repel = TRUE,
  show.expected = TRUE,
  normalization = "subtract",
  label.top = 3,
  label.idents = NULL,
  colors = c("#39B54A", "#5862AD")
)
}
\arguments{
\item{object}{A Seurat object}

\item{assay.1}{Assay containing first set of peaks}

\item{assay.2}{Assay containing second set of peaks}

\item{features}{TF motifs to plot adjacent Tn5 enrichment}

\item{group.by}{Grouping variable to use}

\item{idents}{Identities to group cells by}

\item{label}{Boolean whether to label groups}

\item{repel}{Boolean whether to repel labels}

\item{show.expected}{Boolean whether to plot expected Tn5 integration}

\item{normalization}{Method of normalization of Tn5 integration bias}

\item{label.top}{Number of top groups to label}

\item{label.idents}{Names of identities to label}

\item{colors}{Colors for each plotted CRE set}
}
\value{
ggplot of Tn5 enrichment plots for each motif
}
\description{
This function is modified from Signac's PlotFootprint function.
It plots Tn5 enrichment around inputted motifs in two sets of
peaks, stored in assay.1 and assay.2. Functionality remains
preserved, with the exception of plotting Tn5 enrichment
for two distinct peak sets, returning plots of Tn5 integration
enrichment for each motif
}
