% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_graph_rankings}
\alias{plot_graph_rankings}
\title{Function to generate graph analysis plots}
\usage{
plot_graph_rankings(
  tf_results,
  regulator_tf_names,
  seurat,
  top_n_graph = 20,
  top_n_ranking = 25,
  layout_algorithm = "kk",
  color = "#5862AD"
)
}
\arguments{
\item{tf_results}{List output returned by run_tf_aen with trained model information}

\item{regulator_tf_names}{List of TFs with motif and expression information in dataset}

\item{top_n_graph}{Number of most central TFs to plot on gene regulatory network graph plots}

\item{top_n_ranking}{Number of most central TFs to plot on ranking plots}

\item{layout_algorithm}{Algorithm used to construct gene regulatory network graphs. Options are star, circle, gem, dh, graphopt, grid, mds, randomly, fr, drl, lgl, and kk}

\item{color}{Color of points on ranking plots}
}
\value{
list of plots, two graphs with top TFs by betweenness or PageRank labeled and two ranking plots of top TFs by betweenness or PageRank
}
\description{
This function will produce plots of graph analysis of predicted
gene regulatory networks. It takes in the output of run_tf_aen
and generates graph and ranking plots of top TFs. It calculates
two centrality measures, betweenness and PageRank, to determine
most central TFs to plot. In order to prioritize TFs that are
central, rather than targets, for PageRank, edges are reversed.
The function returns a graph and ranking plot for each measure
of centrality.
}
