% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_motif_enrichment}
\alias{plot_motif_enrichment}
\title{Plot motif enrichment volcano plot}
\usage{
plot_motif_enrichment(
  seurat,
  cres,
  motifs_to_label = NULL,
  num_top_label = 10,
  label_bottom = TRUE,
  num_bottom_label = 5,
  color = "#5862AD",
  flip = FALSE
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{cres}{CRE/peak set to perform motif enrichment on}

\item{motifs_to_label}{Names of motifs to label on plot}

\item{num_top_label}{Number of top motifs to label on plot if motifs_to_label is NULL}

\item{label_bottom}{Boolean to label the least enriched motifs}

\item{num_bottom_label}{If labeling least enriched motifs, number of motifs to label}

\item{color}{Color of points on plot}

\item{flip}{Boolean to flip the x axis. May be useful to plot motif enrichment of two sets of CREs side-by-side}
}
\value{
ggplot of motif enrichment, fold enrichment vs -log(p-value)
}
\description{
This function will plot a one-sided volcano plot of motif
enrichment in the provided CRE/peak set. Axes are fold enrichment
along the x-axis and -log(p-value) along the y-axis. Optional
parameters control labeling of specific motifs. The function returns
a ggplot, in which the x-axis can be flipped if desired, which
may be useful for comparing motif enrichment between two sets of CREs.
}
