% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_tf_rankings}
\alias{plot_tf_rankings}
\title{Plot TFs ordered by regulatory score}
\usage{
plot_tf_rankings(
  results_df,
  tfs_to_label = NULL,
  label_1 = "H",
  label_2 = "FR",
  p_value_cutoff = 0.05,
  score_cutoff = 0.1,
  two_tailed = TRUE,
  top_n_to_label = 5,
  label_tfs = TRUE,
  colors = c("#5862AD", "#39B54A")
)
}
\arguments{
\item{results_df}{Data frame output of rank_tfs, with regulatory score and standard error for each TF}

\item{tfs_to_label}{Names of TFs to label on plot}

\item{label_1}{Name of postive score TFs in plot legend}

\item{label_2}{Name of negative score TFs in plot legend}

\item{p_value_cutoff}{P value cutoff below which TFs are removed from plot. Set to 1 to exclude no TFs}

\item{score_cutoff}{Cutoff for absolute value of regulatory score below which TFs are removed from plot. Set to 0 to exclude no TFs}

\item{two_tailed}{Boolean whether P value cutoff is for a two-tailed test}

\item{top_n_to_label}{Number of top positive and negative TFs to label on plot if tfs_to_label is NULL}

\item{label_tfs}{Boolean whether to label TFs on plot}

\item{colors}{Colors for negative and positive regulatory scores}
}
\value{
ggplot of motif enrichment, fold enrichment vs -log(p-value)
}
\description{
This function inputs the output of rank_tfs to plot TF
by regulatory scores, ordered along the x-axis. TFs with
estimated scores that do not differ from 0 based on
p_value_cutoff or do not meet the score_cutoff threshold
are trimmed to trim the middle of the plot of less relevant
TFs. The function returns a ggplot of ordered rankings.
}
