% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepare_degs}
\alias{prepare_degs}
\title{Select differentially expressed genes for regression}
\usage{
prepare_degs(
  seurat,
  ident.1,
  ident.2,
  group.by = NULL,
  min.pct = 0.1,
  assay = "SCT",
  slot = "data",
  p_val_cutoff = 0.05,
  peak_assay = "peaks",
  ...
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{ident.1}{Character or character vector of cell identities in group 1}

\item{ident.2}{Character or character vector of cell identities in group 2}

\item{group.by}{Name of metadata column to use to group cells by, or NULL if using stored cell identities}

\item{min.pct}{min.pct for FindMarkers}

\item{assay}{Expression assay to use}

\item{slot}{Expression assay slot to use}

\item{p_val_cutoff}{Adjusted p value cutoff to filter DEG list}

\item{peak_assay}{Assay that contains genomic annotation}
}
\value{
List named by inputted gene names, of lists of seeds, each of length num_bootstraps

Differential gene expression data frame
}
\description{
This is a wrapper function for the Seurat FindMarkers function,
taking as input a Seurat object and other parameters to call differentially
expressed genes between two cell groups, then removing genes that are not
in the Seurat's object genome annotation, as CREs to model cannot be selected
without gene coordinates.
}
