% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepare_pseudocell_matrix}
\alias{prepare_pseudocell_matrix}
\title{Generate pseudocell matrix}
\usage{
prepare_pseudocell_matrix(
  seurat,
  assay,
  slot = "data",
  cells_per_partition = 100,
  find_neighbors = FALSE,
  reduction1 = "harmony",
  reduction2 = "harmony_peaks",
  dim_list = list(1:50, 1:50),
  k.nn = 5,
  seed = 489284
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{assay}{Expression or accessibility assay to use}

\item{slot}{Expression or accessibility assay slot to use}

\item{cells_per_partition}{Targeted number of cells per partition}

\item{find_neighbors}{Boolean designating whether to run Seurat's FindMultiModalNeighbors. Set to FALSE if this has already been run}

\item{reduction1}{Reduction 1 to use for FindMultiModalNeighbors, if find_neighbors is TRUE}

\item{reduction2}{Reduction 2 to use for FindMultiModalNeighbors, if find_neighbors is TRUE}

\item{dim_list}{List of dimensions to use for FindMultiModalNeighbors, if find_neighbors is TRUE}

\item{k.nn}{Number of nearest neighbors to use for KNN algorithm, if find_neighbors is TRUE}

\item{seed}{Seed for reproducibility, set to NULL if not needed}
}
\value{
Pseudocell expression/accessibility as a data frame
}
\description{
In order to reduce sparsity and noise, you can elect to generate a pseudocell dataset.
This function takes as input a Seurat object, the desired assay and slot to use for
pseudocell creation, and the number of cells per partition to target. It pools individual
cells into pseudocells with a modified version of VISION's microclustering algorithm to
work on a multiomic WNN graph. Each pseudocell's expression/accessibility is the average of
its constituent individual cells' expression/accessibility. The output is a matrix with
a row for each pseudocell, and column for each gene/peak in the specified assay.
}
