% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{rank_tfs}
\alias{rank_tfs}
\title{Function to rank TFs based on model results}
\usage{
rank_tfs(
  tf_results,
  seurat,
  negative_genes = c(),
  regulator_tf_names = NULL,
  weight_by_expr = TRUE,
  expr_assay = "SCT",
  slot = "data",
  num_cores = 4
)
}
\arguments{
\item{tf_results}{List output returned by run_tf_aen with trained model information}

\item{seurat}{A Seurat object}

\item{negative_genes}{Genes to weight negatively if trying to associate TFs with one of two sides in a trajectory/lineage}

\item{regulator_tf_names}{List of TFs with motif and expression information in dataset}

\item{weight_by_expr}{Boolean to multiply estimated regulatory coefficients by mean TF expression}

\item{expr_assay}{Expression assay to use}

\item{slot}{Assay slot to use}

\item{num_cores}{Number of cores to use for parallelization}
}
\value{
Data frame with regulatory scores for each TF
}
\description{
This function ranks TFs by regulatory influence in the provided
tf_results list. It does this by summing regulatory coefficients
for each TF across all genes. By default, it weights the summed
coefficients by the mean expression of the TF, following the
hypothesis that highly expressed TFs are more likely to be
relevant than lowly expressed TFs, so that the computed regulatory
score is effectively the number of target transcripts increased by
one unit of expression of the given TF. Optionally, the names of
genes to weight negatively can be provided in order to weight
regulatory coefficients for each modeled gene. This can be helpful
when ranking TFs predicted to regulate a set of differentially
expressed genes in order to predict TFs that are associated or
potentially drive either side of the comparison.
}
