% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{rank_tfs_by_centrality}
\alias{rank_tfs_by_centrality}
\title{Function to rank TFs based on graph analysis of model results}
\usage{
rank_tfs_by_centrality(tf_results, seurat, assay = "SCT", slot = "data")
}
\arguments{
\item{tf_results}{List output returned by run_tf_aen with trained model information}
}
\value{
List of two rankings, one by betweenness and one by PageRank
}
\description{
This function ranks TFs by centrality in the provided
tf_results list. It does this by constructing gene regulatory
networks using the estimated coefficients of each TF for
each gene. The function returns two rankings by centrality:
betweenness and PageRank. In order to prioritize TFs that are
central, rather than targets, for PageRank, edges are reversed.
}
