% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{run_tf_aen_ccan}
\alias{run_tf_aen_ccan}
\title{Run adaptive elastic net to identify TFs regulating target genes with paired, non-multiome single cell datasets}
\usage{
run_tf_aen_ccan(
  seurat,
  pseudocell_mat,
  motif_matrix_df,
  peak_gene_key,
  gene_list = NULL,
  lambda2 = 0.5,
  gamma = 1,
  ci_cutoff = 1.96,
  pval_cutoff = NULL,
  set_nonsig_to_zero = TRUE,
  promoter_only = FALSE,
  promoter_threshold = 2000,
  train_fraction = 0.8,
  num_bootstraps = 100,
  bootstrap = TRUE,
  num_threads = 16,
  mc_cores = 8,
  globals_maxsize = NULL,
  verbose = TRUE,
  bs_seed = 943126,
  peak_assay = "peaks",
  multi_seed = 6283408,
  ...
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{pseudocell_mat}{Pseudocell expression matrix to use}

\item{motif_matrix_df}{motif-peak dataframe returned from running make_motif_matrix on a scATAC-seq Seurat object and a PWM list of motifs to use}

\item{peak_gene_key}{peak-gene dataframe returned from running conns_to_peak_gene_key on the connections generated by Cicero}

\item{gene_list}{Names of genes to model}

\item{lambda2}{Parameter for quadratic L2 penalty for regularization}

\item{gamma}{Positive constant for calculating adaptive weights}

\item{ci_cutoff}{Confidence interval cutoff used to determine whether an estimated non-zero coefficient should be kept}

\item{pval_cutoff}{As an alternative, p value cutoff to use in order to calculate ci_cutoff. E.g. p value of 0.5 corresponds to 95\% CI.}

\item{set_nonsig_to_zero}{Boolean to determine whether non-significant coefficient estimates should be set to zero}

\item{promoter_only}{Boolean determining whether modeling is run on TFs with motifs present in gene promoter regions only}

\item{promoter_threshold}{Distance from TSS to determine promoter region to classify overlapping peaks in dataset as promoter peaks}

\item{train_fraction}{Fraction of pseudocells used for training during cross-validation step}

\item{num_bootstraps}{Number of bootstrap samples}

\item{bootstrap}{Boolean if set to false, will not perform bootstrapping}

\item{num_threads}{Number of threads to use for multithreading}

\item{globals_maxsize}{If running into RAM problems during multithreading, can use this parameter to set future.globals.maxSize if not done manually}

\item{verbose}{Boolean to determine whether to run silently}

\item{bs_seed}{Seed for reproducibility in determining bootstrap seeds, set to NULL if not needed}

\item{multi_seed}{Seed for reproducibility during multithreaded model training, set to NULL if not needed}
}
\value{
list of lists per modeled gene, containing modeling and TF gene regulatory network information
}
\description{
This function will identify putative transcription factors regulating each of
the inputted genes. It takes as input a single cell RNA-seq Seurat object, the pseudocell expression
matrix generated by prepare_pseudocell_matrix, the motif-peak dataframe generated
by make_motif_matrix, the peak-gene dataframe generated by conns_to_peak_gene_key,
and parameters for running the adaptive elastic net and parallelization.
The function returns information for each modeled gene as a list. Each gene's entry in
this list is a list containing the following information: the coefficient of determination
calculated during the cross validation step with the calculated lambda.min value,
the trained model object, and a data frame with regulatory weights and standard
errors for all statistically significant TFs. Regulatory weights for
non-significant TFs are set to 0 for future analysis if desired.
}
