% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{UpstreamDistanceToTSS}
\alias{UpstreamDistanceToTSS}
\title{Utility function taken from Signac DistanceToTSS function, for considering only upstream promoter region for peaks}
\usage{
UpstreamDistanceToTSS(peaks, genes, distance = 2e+05, sep = c("-", "-"))
}
\description{
This function is a copy of the Signac DistanceToTSS function, except for
changing the downstream parameter in the Extend function call from distance
to 0.
}
